requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-guardian==2.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==6.1.0
psycopg2-binary==2.8.4
Pygments==2.4.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.3
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.2.3
attrs==19.3.0
autopep8==1.5
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2019.6.16
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cupshelpers==1.0
defer==1.0.6
Django==3.0.3
django-filter==2.2.0
django-guardian==2.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
httplib2==0.9.2
idna==2.8
importlib-metadata==1.5.0
itypes==1.1.0
Jinja2==2.10.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.2.0
netifaces==0.10.4
onboard==1.4.1
packaging==20.1
PAM==0.4.2
pexpect==4.2.1
Pillow==6.1.0
pluggy==0.13.1
protobuf==3.0.0
psutil==5.4.2
py==1.8.1
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.4.2
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.6
pyparted==3.11.1
pyRFC3339==1.0
pytest==5.3.5
python-apt==1.6.4
python-debian==0.1.32
python-xapp==1.8.1
python-xlib==0.20
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.14.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tinycss==0.4
ubuntu-drivers-common==0.0.0
ufw==0.36
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==15.1.0
wcwidth==0.1.8
whitenoise==4.1.3
xkit==0.0.0
zipp==3.0.0
